Try to fix bugs which prevent compiling the nigly builds.
I have no clang but it is a shame that gcc did not find the problems.

- the clang stuff shall be fixed
- most of the icc stuff shall be fixed but not the shared pointer stuff yet
- the mingw stuff could work now



82% tests passed, 17 tests failed out of 95

Total Test time (real) =  61.37 sec

The following tests FAILED:
	  5 - Singlefield_Acoustics_PmlVarDensity2d (Failed)
	 14 - Singlefield_Acoustics_Mortar2d (Failed)
	 27 - Singlefield_Heat_HeatMortar (Failed)
	 44 - Singlefield_Magnetics_Coil2dMortar (Failed)
	 51 - Singlefield_Mechanics_PreStress3dEV (Failed)
	 60 - Singlefield_Mechanics_dampedEV2D (Failed)
	 61 - Singlefield_Mechanics_BlochMode2D (Failed)
	 62 - Singlefield_Mechanics_BlochMode3D (Failed)
	 68 - Coupledfield_MechAcou_RingHarm2d (Failed)
	 78 - Coupledfield_MagMech_PermaMSRestart (Failed)
	 79 - Coupledfield_MagMechAcou_PlateAboveCoilFluid2D (Failed)
	 86 - Optimization_Homogenization_homogenize_only (Failed)
	 88 - Optimization_Dynamic_eigenfrequency (Failed)
	 89 - Optimization_Dynamic_bloch2d (Failed)
	 92 - Optimization_FMO_mech_coef_ev (Failed)
	 93 - Optimization_ParamMat_TopMat_DensityTimesRotatedTransversalIsotropic (Failed)
	 94 - Optimization_ParamMat_TopMat_DensityTimesRotatedInterpolatedTensor (Failed)

+++ source/Optimization/Optimization.cc
@@ -105,7 +105,7 @@ Optimization::Optimization()
 
 
   if(driver->GetDriverClass() != BaseDriver::SINGLE_DRIVER)
-    throw Exception("optimization not implemented for driver " + driver->GetDriverClass());
+    throw Exception("optimization not implemented for driver " + (int) driver->GetDriverClass());
 
   optInfoNode = domain->GetInfoRoot()->Get("optimization");   // store our info results here
   PtrParamNode header = optInfoNode->Get(ParamNode::HEADER);
